T4T NDEF detection now working with sizes > 256 bytes

+++ rpntts-backend/rpntts-common.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "rpntts-common.h"
 
-void bin_to_hex(uint8_t *in, uint8_t in_length, char *out) {
+void bin_to_hex(uint8_t *in, uint16_t in_length, char *out) {
     char *ppos;
     uint16_t i;
